+++ src/main/java/ru/kunik/uct/numerator/NumeratorHandler.java
@@ -6,7 +6,7 @@ import ru.kunik.uct.MainHandler;
 
 public class NumeratorHandler extends MainHandler {
     
-    public static volatile List<Integer> numeratorList = new ArrayList<Integer>();
+    public static volatile List<Long> numeratorList = new ArrayList<Long>();
     
     public NumeratorHandler() {
         super("1.0.0", "Numerator");
